Fixes for #3 | Update element/model/page/generators
Modified generator to convert name to camelCase for function or object names
Still configured to be Using kebab-case for file names
Also modified elements and pages to use function names for easier to read stack traces.
Modified model to have a name as well.

Elements are using bel instead of choo/html