adding a check for installation of yarn
the check is pretty simple, it makes sure that if a yarn.lock file is in the cwd that's being installed to, yarn is being used for installation.

I'm not 100% sure what `npm install -s` is supposed to do differently than `npm install`, and I couldn't find an alternative for yarn, so I kept it at this.

Let me know if I can do anything better or another way

Have a nice day!